+++ specific/registry.h
@@ -52,6 +52,7 @@
 #define REG_INVBGND_MODE		"InvBackgroundMode"
 #define REG_PICTURE_STRETCH		"PictureStretchLimit"
 #define REG_LOADING_SCREENS		"LoadingScreens"
+#define REG_CUSTOM_WATER_COLOR	"CustomWaterColor"
 #define REG_SHADOW_MODE			"ShadowMode"
 #define REG_REFLECTION_MODE		"ReflectionMode"
 #define REG_REFLECTION_BLUR		"ReflectionBlur"