+++ specific/smain.cpp
@@ -65,6 +65,7 @@ extern DWORD ShadowMode;
 extern DWORD AlphaBlendMode;
 extern DWORD ReflectionMode;
 extern DWORD ReflectionBlur;
+extern bool CustomWaterColorEnabled;
 #endif // FEATURE_VIDEOFX_IMPROVED
 
 #ifdef FEATURE_SCREENSHOT_IMPROVED
@@ -498,6 +499,7 @@ void __cdecl S_LoadSettings() {
 	GetRegistryDwordValue(REG_ALPHABLEND_MODE, &AlphaBlendMode, 0);
 	GetRegistryDwordValue(REG_REFLECTION_MODE, &ReflectionMode, 0);
 	GetRegistryDwordValue(REG_REFLECTION_BLUR, &ReflectionBlur, 2);
+	GetRegistryBoolValue(REG_CUSTOM_WATER_COLOR, &CustomWaterColorEnabled, false);
 	CLAMPG(AlphaBlendMode, 2);
 	CLAMPG(ReflectionMode, 2);
 	CLAMPG(ReflectionBlur, 7);